Removing `eval`
Using `eval` with user input is dangerous as it opens to door to
code execution.

It was used in 3 types of use cases:
1) transforming hex values to integers. So just cast them to an int
2) to map page sizes to constants. So made a map and defaulted to A4
3) for interpreting styles. Normally these are weldefined key value
type of strings which are evaluated to a dict. So using
`ast.literal_eval` to do this in a safe way.

In the 3 point there was catch as user input could wrongfully be
considered as a style if it had the "right" prefix.